feat!: Remove trait MaxNumDigits from public interface
It is now an implementation detail of integer
