+++ src/integer.rs
@@ -7,7 +7,7 @@ use num_traits::{Bounded, CheckedAdd, CheckedMul, CheckedSub, One, Zero};
 
 use crate::{
     FromDigit, FromRadix10, FromRadix10Checked, FromRadix10Signed, FromRadix10SignedChecked,
-    FromRadix16, FromRadix16Checked, MaxNumDigits, Sign,
+    FromRadix16, FromRadix16Checked, Sign,
 };
 
 /// Wrapper which implements the traits [`crate::FromRadix10`], [`crate::FromRadix10Checked`],
@@ -291,6 +291,18 @@ where
     }
 }
 
+/// A bounded integer, whose representation can overflow and therefore can only store a maximum
+/// number of digits
+trait MaxNumDigits {
+    /// Given a representation with a radix character I, what is the maximum number of digits we can
+    /// parse without the integer overflowing for sure?
+    fn max_num_digits(radix: Self) -> usize;
+
+    /// Returns the maximum number of digits a negative representation of `I` can have depending on
+    /// `radix`.
+    fn max_num_digits_negative(radix: Self) -> usize;
+}
+
 impl<I> MaxNumDigits for Integer<I>
 where
     I: Bounded + Zero + DivAssign + Ord + Copy,
@@ -317,5 +329,30 @@ where
             min /= radix.0;
         }
         d
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn max_digits() {
+        assert_eq!(10, Integer::<i32>::max_num_digits(Integer(10)));
+        assert_eq!(10, Integer::<u32>::max_num_digits(Integer(10)));
+        assert_eq!(19, Integer::<i64>::max_num_digits(Integer(10)));
+        assert_eq!(20, Integer::<u64>::max_num_digits(Integer(10)));
+        assert_eq!(3, Integer::<u8>::max_num_digits(Integer(10)));
+        assert_eq!(3, Integer::<i8>::max_num_digits(Integer(10)));
+    }
+
+    #[test]
+    fn max_digits_negative() {
+        assert_eq!(10, Integer::<i32>::max_num_digits_negative(Integer(10)));
+        assert_eq!(0, Integer::<u32>::max_num_digits_negative(Integer(10)));
+        assert_eq!(19, Integer::<i64>::max_num_digits_negative(Integer(10)));
+        assert_eq!(0, Integer::<u64>::max_num_digits_negative(Integer(10)));
+        assert_eq!(0, Integer::<u8>::max_num_digits_negative(Integer(10)));
+        assert_eq!(3, Integer::<i8>::max_num_digits_negative(Integer(10)));
     }
 }
